Allow for iterable match config
The `Str::is()` method is used for testing the matches against URIs and route names; `Str::is($match)` is allowed to be either a string or an iterable. By modifying the LaravelApiExplorer::filterRoutes() method to filter the config `laravelapiexplorer.match` as if it could be a string or an iterable, we thereby allow for using an iterable in the configuration. This allows for more control over which routes appear in the explorer, without having to use the ignore configuration.